Moved FFI layer to a new crate.
To allow using meshlite as a Rust library while preserving the FFI layer as is,
with the exception of the name change to `meshlite_ffi`.
