+++ ffi/README.md
@@ -1,0 +1,3 @@
+
+C ABI layer for meshlite. Allows building meshlite as a dynamic library.
+
