+++ ffi/appveyor.yml
@@ -24,8 +24,8 @@ test_script:
 
 after_test:
   - 7z a meshlite_unstable_vc14_%PLATFORM%.zip %APPVEYOR_BUILD_FOLDER%\include\meshlite.h
-  - 7z a meshlite_unstable_vc14_%PLATFORM%.zip %APPVEYOR_BUILD_FOLDER%\target\release\meshlite.dll
-  - 7z a meshlite_unstable_vc14_%PLATFORM%.zip %APPVEYOR_BUILD_FOLDER%\target\release\meshlite.dll.lib
+  - 7z a meshlite_unstable_vc14_%PLATFORM%.zip %APPVEYOR_BUILD_FOLDER%\target\release\meshlite_ffi.dll
+  - 7z a meshlite_unstable_vc14_%PLATFORM%.zip %APPVEYOR_BUILD_FOLDER%\target\release\meshlite_ffi.dll.lib
 
 artifacts:
   - path: '*.zip'
